chore(ci): clear clippy gating + safe npm audit (subset of inherited CI debt)
Forensic audit of GitHub Actions runs on this branch vs main shows all 4
red checks (no-hardcode, Rust CI, Security Scans, unit-tests) pre-existed
on main (commit 2da6d82c) before B1.d was pushed. e2e is green.

This commit clears the gating subset that traces back to commits I
authored in prior sprints:

CLIPPY GATING (3 errors, all in code I introduced earlier):

- backend/relays-client/src/submit_engine.rs:105 — OMEGA-6 commit
  e601a411 cast `opp.chain_id as u64` when the field is already u64.
  `clippy::unnecessary_cast` fires on `-D warnings`. Fix: drop the cast.

- backend/searcher-rs/src/counters.rs:320 — B0 commit 0244f122 used
  `REGISTRY.write().expect(...)`; lib.rs has
  `#![deny(clippy::expect_used)]`. Fix: explicit match that recovers
  from poison via `into_inner()` (counters are atomic primitives, so
  a poisoned guard is recoverable — documented inline).

- backend/searcher-rs/src/scanner.rs:1010 — db36a88 (50ms timeout
  fix) wrote `Err(e.into())` where the From conversion is identity.
  `clippy::useless_conversion` fires on `-D warnings`. Fix: drop
  `.into()`.

Cargo clippy -p searcher-rs -p relays-client -- -D warnings now CLEAN.
Cargo test -p searcher-rs --lib counters: 6/6 PASS — no regression.

NPM AUDIT (prod deps, high+) — safe path:

`npm audit fix --omit=dev` (no --force) bumped a transitive postcss-
adjacent dep. Local `npm audit --omit=dev --audit-level=high` now
EXIT=0. Remaining vulns are dev-only (filtered by --omit=dev) and
require Next.js 16 major bump to clear — deferred to a dedicated
upgrade sprint with frontend regression validation.

DEFERRED (NOT IN THIS COMMIT — scope rationale per category):

- no-hardcode lint (259 violations): zero from my B1.d push;
  pre-existing across triangular_worker.rs, price_oracle.rs,
  credentials/validators.ts, erc20_storage.rs, etc. Doctrinal
  cleanup spanning 30+ files — separate sprint with per-category
  triage (canonical tokens → chains.rs constants; external URLs →
  env vars; test fixtures → fixture allowlist).

- cargo audit Rust advisories: requires `cargo install cargo-audit`
  locally + workspace-pinned bumps of redis 0.25 / sqlx-postgres 0.7.4
  / etc. API-breaking risk across api-server + searcher-rs + relays-
  client. Separate sprint with per-crate compatibility matrix.

- TS unit-tests testcontainer failure: api-server integration tests
  (test/migrations.test.ts, test/opportunities-live.test.ts) require
  a Docker daemon for testcontainers. Failure mode is infra, not
  code. Either CI runner needs Docker pre-configured, or these
  tests need to be marked as integration and excluded from the
  `--lib` CI gate (current Rust CI workflow already does this for
  Rust side via `--lib`; equivalent TS gate is the open item).

+++ backend/relays-client/src/submit_engine.rs
@@ -102,7 +102,7 @@ impl SubmitEngine {
         // never affects Chain Y. Falls back to legacy `arbx:papermode` for 30
         // days from 2026-05-13 (PaperModeClient::state_for_chain handles the
         // fallback chain internally with PaperModeSource attribution).
-        let paper_dynamic = self.paper_mode.is_enabled_for_chain(opp.chain_id as u64).await;
+        let paper_dynamic = self.paper_mode.is_enabled_for_chain(opp.chain_id).await;
         let paper = paper_dynamic || paper_env;
 
         // -----------------------------------------------------------------------
